+++ src/main/java/com/umc/smupool/SmupoolApplication.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
 @EnableJpaAuditing
-@EntityScan(basePackages = "com.umc.smupool.domain.member.entity")
 public class SmupoolApplication {
 
 	public static void main(String[] args) {
