+++ src/main/java/com/umc/smupool/global/Security/exception/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.umc.smupool.global.config.Security.jwt;
+package com.umc.smupool.global.Security.exception;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.umc.smupool.global.apiPayload.ApiResponse;
@@ -6,21 +6,22 @@ import com.umc.smupool.global.apiPayload.code.status.GeneralErrorCode;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
 @Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
     @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
+    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
         response.setContentType("application/json; charset=UTF-8");
-        response.setStatus(403);
+        response.setStatus(401);
 
         ApiResponse<Object> errorResponse = ApiResponse.onFailure(
-                GeneralErrorCode.FORBIDDEN_403.getCode(), GeneralErrorCode.FORBIDDEN_403.getMessage(), null);
+                GeneralErrorCode.UNAUTHORIZED_401.getCode(), GeneralErrorCode.UNAUTHORIZED_401.getMessage(), null);
 
         ObjectMapper mapper = new ObjectMapper();
         mapper.writeValue(response.getOutputStream(), errorResponse);
