+++ src/main/java/com/umc/smupool/global/Security/util/JWTUtil.java
@@ -1,4 +1,4 @@
-package com.umc.smupool.global.config.Security.jwt;
+package com.umc.smupool.global.Security.util;
 
 import com.umc.smupool.domain.member.exception.AuthErrorCode;
 import com.umc.smupool.domain.member.exception.handler.AuthHandler;
